% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca.R
\name{clean_species_name}
\alias{clean_species_name}
\title{Clean species names encoded in source filenames}
\usage{
clean_species_name(x)
}
\arguments{
\item{x}{Character vector such as
`Sorted_aglais_urticae_broad_High_distances`.}
}
\value{
Cleaned binomial names.
}
\description{
Clean species names encoded in source filenames
}
