% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{default_fpca_schemes}
\alias{default_fpca_schemes}
\title{Default habitat-definition scenarios}
\usage{
default_fpca_schemes()
}
\value{
A character vector containing the six scenarios used in the
  manuscript.
}
\description{
Default habitat-definition scenarios
}
